Add resource that lists all locations combined
CO-259

This resource gathers locations from campusmap, extension and dining.
The information is JSONAPI format and written to a single file.

This endpoint has been used for testing. The different datasources
are changed at different rates, so it might be worthwhile separating
the json data into multiple json files to be uploaded to ES.