Test that we are constructing ArcGisLocations correctly
CO-433

If you pass the wrong params to ArcGisLocation, it can silently
construct an object with all null parameters, which would cause false
negatives in the test.

This bit me when writing the tests, so add a little assertion to one of
the tests that makes sure ArcGisLocation was constructed correctly.