Fix overriding of recurring events
ECSPCS-311

When a calendar contains a recurring event for which at least one
instance has been modified, we would sometimes use the outdated event
instead of the updated one.

A modified instance of a recurring event is represented in an iCalendar
by a second VEVENT structure which has the same UID as the original
recurring event, and a RECURRENCE-ID equal to the start time of the
event to be replaced.

We already have logic to drop events with duplicate UIDs from the dining
results when one event has a higher sequence number than the other (which is
the other way you can wind up with duplicate UIDs) but it is normal for
recurring events to share the same sequence number, and when faced with
this the code threw up its hands and fell back on preferring the event
which occured lexicographically later in the ics file, which happened to
be the wrong choice.

This commit enhances that logic to prefer an event which has a
RECURRENCE-ID to one that does not, then the event with the highest
SEQUENCE number, then LAST-MODIFIED if we still aren't sure (on the
theory that most recently modified event is more likely to be the right
one).

A more proper approach would actually examine the RECURRENCE-ID,
calculate the recurrence set of the original event, and only those
events in the set with the same start time, but since we only examine a
single day's events at a time and the calendars we are interested in
don't repeat events any more frequently than daily, we can get a way
with the simpler approach.

The iCalendar spec[1] isn't clear on exactly how RECURRENCE-ID is
supposed to interact with SEQUENCE, so we're forced to just make a best
guess. There is some speculation at [2], if anyone is interested in
making a more rigorous attempt.

[1]: https://www.ietf.org/rfc/rfc2445.txt
[2]: http://icalevents.com/4437-correct-handling-of-uid-recurrence-id-sequence/