Work around "duplicate thursdays" bug
ECSPCS-311

The locations api would sometimes return multiple, overlapping dining
hours for a location in a single day. For example, it would say that
Java II is open on Thursday from 7:30am to midnight, and also from
7:30am to 11pm.

Turns out that the 7:30am-midnight time is actually for the Wednesday
before, but because we ask for all events between midnight and midnight
on Thursday, ical4j erroneously considers the Wednesday time (which ends
precisely at midnight) to intersect with our requested period.

The workaround is to use a period of one second after midnight to one
second before the following midnight. Note that an event which goes
past midnight will still match both days. A more thorough fix would
probably be to only associate events which *start* on a day with that
day.