fix library open hours bug
CO-942

MergeUtil.populate() was looking for a location with bldgID of VLib to
add the hours to. But the library's bldgID is "0036"; its abbreviation
is "VLib". MergeUtil used to check abbreviation, but that got
accidentally changed in CO-926, which changed most of the code to use
bldgID rather than abbreviation.