fix timezone issues in DiningDAOTest
DiningDAOTest checks that IcalUtil.getEventsForDay returns the correct
days by comparing against hardcoded dates. But the hardcoded dates are
constructed without an explicit timezone, so they default to the system
time. This meant that the test would fail if the system timezone was not
set to US Pacific time.

Fix this by setting an explicit timezone.