WIP unit tests for Cache
CO-955

I don't think this even compiles right now.

My line in the sand is that a test which hits the network is not a unit
test, so my main concern here is trying to isolate Cache from the
network during testing, which is a little difficult because half the
point of the class is dealing with network errors.

You could try to isolate from the filesystem too, but it probably isn't
worth it.

The main method to test is withDataFromUrlOrCache.

I started three different approaches to doing this and i'm not sure
which one is the best.
* creating a MockCache class which replaces getURL and would maybe would
    have methods for setting the contents that getURL method returns.
* the same, but using anonymous classes to construct the behaviour
    inline in the test that the Cache is used in.
* using mockito. actually, i'm currently using mockito in conjuction
    with option #2 above, but maybe it could do the whole job?

I'm not sure the getDataFromCache/writeDataToCache methods are worth
testing.

Good luck.